Move application status functions to their own file
We will be using these in a subsequent commit. If they are using in their original file it will
create a dependency cycle and given as the 'utils' file is quite long it makes sense to split
them up into their own file.
