+++ server/utils/applications/utils.test.ts
@@ -1,4 +1,4 @@
-import { ApplicationSortField, ApprovedPremisesApplicationStatus } from '@approved-premises/api'
+import { ApplicationSortField } from '@approved-premises/api'
 import { isAfter } from 'date-fns'
 import { mockOptionalQuestionResponse } from '../../testutils/mockQuestionResponse'
 import {
@@ -23,7 +23,6 @@ import { isApplicableTier, isFullPerson, nameOrPlaceholderCopy, tierBadge } from
 import {
   appealDecisionRadioItems,
   applicationStatusSelectOptions,
-  applicationStatuses,
   applicationTableRows,
   dashboardTableHeader,
   dashboardTableRows,
@@ -36,7 +35,6 @@ import {
   lengthOfStayForUI,
   mapPlacementApplicationToSummaryCards,
   mapTimelineEventsForUi,
-  statusTags,
   withdrawCell,
   withdrawnStatusTag,
 } from './utils'
@@ -454,37 +452,6 @@ describe('utils', () => {
             },
           ],
         ])
-      })
-    })
-  })
-
-  describe('statusTags', () => {
-    it('should return a list of tags for each status', () => {
-      expect(statusTags()).toEqual({
-        assesmentInProgress: '<strong class="govuk-tag govuk-tag--blue">Assessment in progress</strong>',
-        awaitingAssesment: '<strong class="govuk-tag govuk-tag--blue">Awaiting assessment</strong>',
-        awaitingPlacement: '<strong class="govuk-tag govuk-tag--blue">Awaiting placement</strong>',
-        inapplicable: '<strong class="govuk-tag govuk-tag--red">Application inapplicable</strong>',
-        placementAllocated: '<strong class="govuk-tag govuk-tag--pink">Placement allocated</strong>',
-        rejected: '<strong class="govuk-tag govuk-tag--red">Application rejected</strong>',
-        requestedFurtherInformation:
-          '<strong class="govuk-tag govuk-tag--yellow">Further information requested</strong>',
-        started: '<strong class="govuk-tag govuk-tag--blue">Application started</strong>',
-        submitted: '<strong class="govuk-tag govuk-tag--">Application submitted</strong>',
-        unallocatedAssesment: '<strong class="govuk-tag govuk-tag--blue">Unallocated assessment</strong>',
-        withdrawn: '<strong class="govuk-tag govuk-tag--red">Application withdrawn</strong>',
-        pendingPlacementRequest: '<strong class="govuk-tag govuk-tag--blue">Pending placement request</strong>',
-      })
-    })
-  })
-
-  describe('getStatus', () => {
-    const statuses = Object.keys(applicationStatuses) as Array<ApprovedPremisesApplicationStatus>
-
-    statuses.forEach(status => {
-      it(`returns the correct tag for each an application with a status of ${status}`, () => {
-        const application = applicationSummaryFactory.build({ status })
-        expect(getStatus(application)).toEqual(statusTags()[status])
       })
     })
   })
